requirements.txt
dash
pandas
python-gitlab
